Added the README and also the screenshot of the sample run that I ran to demonstrate it running.
