+++ WebCrawler/WebCrawler/Debug/WebCrawler.log
@@ -1,3 +1,3 @@
-﻿  main.cpp
+﻿  weiler-conrad-gore-main.cpp
 LINK : warning LNK4067: ambiguous entry point; selected 'mainCRTStartup'
   WebCrawler.vcxproj -> C:\Users\Matthew\Dropbox\University-Of-Dayton\Spring2018Masters\Data-Communications\Assignment1\WebCrawler\Debug\WebCrawler.exe
