+++ WebCrawler/WebCrawler/WebCrawler.vcxproj
@@ -83,11 +83,11 @@
     <ClInclude Include="common.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
-    <ClInclude Include="urlparser.h" />
-    <ClInclude Include="winsock.h" />
+    <ClInclude Include="weiler-conrad-gore-urlparser.h" />
+    <ClInclude Include="weiler-conrad-gore-winsock.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="main.cpp" />
+    <ClCompile Include="weiler-conrad-gore-main.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
